+++ plotnine/mapping/aes.py
@@ -1,16 +1,22 @@
 from __future__ import annotations
 
 import re
+import typing
 from collections.abc import Iterable
 from contextlib import suppress
 from copy import deepcopy
 from dataclasses import fields
-from typing import Any, Dict, overload
+from typing import Any, Dict
 
 import pandas as pd
 
 from ..iapi import labels_view
 from .evaluation import after_stat, stage
+
+if typing.TYPE_CHECKING:
+    from typing import TypeVar
+
+    THasAesNames = TypeVar("THasAesNames", bound=list[str] | dict[str, Any])
 
 __all__ = ("aes",)
 
@@ -186,6 +192,9 @@ class aes(Dict[str, Any]):
         kwargs = self._convert_deprecated_expr(kwargs)
         self.update(kwargs)
 
+    def __iter__(self):
+        return iter(self.keys())
+
     def _convert_deprecated_expr(self, kwargs):
         """
         Handle old-style calculated aesthetic expression mappings
@@ -202,7 +211,7 @@ class aes(Dict[str, Any]):
         return kwargs
 
     @property
-    def _starting(self):
+    def _starting(self) -> dict[str, Any]:
         """
         Return the subset of aesthetics mapped from the layer data
 
@@ -219,7 +228,7 @@ class aes(Dict[str, Any]):
         return d
 
     @property
-    def _calculated(self):
+    def _calculated(self) -> dict[str, Any]:
         """
         Return only the aesthetics mapped to calculated statistics
 
@@ -234,7 +243,7 @@ class aes(Dict[str, Any]):
         return d
 
     @property
-    def _scaled(self):
+    def _scaled(self) -> dict[str, Any]:
         """
         Return only the aesthetics mapped to after scaling
 
@@ -295,19 +304,7 @@ class aes(Dict[str, Any]):
         return new
 
 
-@overload
-def rename_aesthetics(obj: list[str]) -> list[str]:
-    ...
-
-
-@overload
-def rename_aesthetics(obj: dict[str, Any]) -> dict[str, Any]:
-    ...
-
-
-def rename_aesthetics(
-    obj: list[str] | dict[str, Any]
-) -> list[str] | dict[str, Any]:
+def rename_aesthetics(obj: THasAesNames) -> THasAesNames:
     """
     Rename aesthetics in obj
 
@@ -327,7 +324,9 @@ def rename_aesthetics(
             if name != new_name:
                 obj[new_name] = obj.pop(name)
     else:
-        obj = [name.replace("colour", "color") for name in obj]
+        for i, name in enumerate(obj):
+            if "colour" in name:
+                obj[i] = name.replace("colour", "color")
 
     return obj