+++ plotnine/stats/stat_summary_bin.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 from ..doctools import document
 from ..exceptions import PlotnineWarning
-from ..scales.scale import scale_discrete
+from ..scales.scale_discrete import scale_discrete
 from ..utils import groupby_apply
 from .binning import fuzzybreaks
 from .stat import stat
@@ -141,9 +141,9 @@ class stat_summary_bin(stat):
 
         breaks = fuzzybreaks(scales.x, breaks, boundary, binwidth, bins)
         data["bin"] = pd.cut(
-            data["x"],  # pyright: ignore
+            data["x"],
             bins=breaks,  # pyright: ignore
-            labels=False,  # pyright: ignore
+            labels=False,
             include_lowest=True,
         )