+++ plotnine/utils.py
@@ -3,11 +3,11 @@ Little functions used all over the codebase
 """
 from __future__ import annotations
 
-import collections
 import inspect
 import itertools
 import typing
 import warnings
+from collections import defaultdict
 from contextlib import suppress
 from warnings import warn
 from weakref import WeakValueDictionary
@@ -29,7 +29,13 @@ if typing.TYPE_CHECKING:
     from IPython.core.interactiveshell import InteractiveShell
     from typing_extensions import TypeGuard
 
-    from plotnine.typing import DataLike, FloatArray, FloatArrayLike, IntArray
+    from plotnine.typing import (
+        AnyArrayLike,
+        DataLike,
+        FloatArray,
+        FloatArrayLike,
+        IntArray,
+    )
 
 
 # Points and lines of equal size should give the
@@ -109,6 +115,37 @@ def match(
     else:
         lst = [lookup[x] + start if x in lookup else nomatch for x in v1]
     return np.array(lst)
+
+
+def multitype_sort(arr: AnyArrayLike) -> list[Any]:
+    """
+    Sort elements of multiple types
+
+    x is assumed to contain elements of different types, such that
+    plain sort would raise a `TypeError`.
+
+    Parameters
+    ----------
+    a : array-like
+        Array of items to be sorted
+
+    Returns
+    -------
+    out : list
+        Items sorted within their type groups.
+    """
+    types = defaultdict(list)
+
+    for x in arr:
+        if isinstance(x, (int, float, complex)):
+            types["number"].append(x)
+        else:
+            types[type(x)].append(x)
+
+    for t, values in types.items():
+        types[t] = sorted(values)
+
+    return list(itertools.chain.from_iterable(types[t] for t in types))
 
 
 def _margins(
@@ -262,7 +299,7 @@ def ninteraction(df: pd.DataFrame, drop: bool = False) -> list[int]:
 
     combs = np.array(np.hstack([1, np.cumprod(ndistinct[:-1])]))
     mat = np.array(ids)
-    res = (mat - 1) @ combs.T + 1
+    res = (mat - 1) @ combs.T + 1  # type: ignore
     res = np.array(res).flatten().tolist()
 
     if drop:
@@ -297,7 +334,7 @@ def _id_var(x: pd.Series[Any], drop: bool = False) -> list[int]:
             if has_nan:
                 # NaNs are -1, we give them the highest code
                 nan_code = -1
-                new_nan_code = np.max(x.cat.codes) + 1
+                new_nan_code = np.max(x.cat.codes) + 1  # type: ignore
                 lst = [val if val != nan_code else new_nan_code for val in x]
             else:
                 lst = list(x.cat.codes + 1)
@@ -305,10 +342,8 @@ def _id_var(x: pd.Series[Any], drop: bool = False) -> list[int]:
         try:
             levels = sorted(set(x))
         except TypeError:
-            from mizani.utils import multitype_sort
-
             # x probably has NANs
-            levels = multitype_sort(set(x))
+            levels = multitype_sort(list(set(x)))
 
         lst = match(x, levels)
         lst = [item + 1 for item in lst]
@@ -358,16 +393,16 @@ def check_required_aesthetics(required, present, name):
         raise PlotnineError(msg.format(name, ", ".join(missing_aes)))
 
 
-def uniquecols(df):
+def uniquecols(data):
     """
     Return unique columns
 
     This is used for figuring out which columns are
     constant within a group
     """
-    bool_idx = df.apply(lambda col: len(np.unique(col)) == 1, axis=0)
-    df = df.loc[:, bool_idx].iloc[0:1, :].reset_index(drop=True)
-    return df
+    bool_idx = data.apply(lambda col: len(np.unique(col)) == 1, axis=0)
+    data = data.loc[:, bool_idx].iloc[0:1, :].reset_index(drop=True)
+    return data
 
 
 def jitter(x, factor=1, amount=None, random_state=None):
@@ -427,16 +462,16 @@ def jitter(x, factor=1, amount=None, random_state=None):
         elif xx != 0:
             d = xx / 10.0
         else:
-            d = z / 10
+            d = z / 10  # type: ignore
         amount = factor / 5.0 * abs(d)
     elif amount == 0:
-        amount = factor * (z / 50.0)
+        amount = factor * (z / 50.0)  # type: ignore
 
     return x + random_state.uniform(-amount, amount, len(x))
 
 
 def remove_missing(
-    df: pd.DataFrame,
+    data: pd.DataFrame,
     na_rm: bool = False,
     vars: Sequence[str] | None = None,
     name: str = "",
@@ -457,27 +492,29 @@ def remove_missing(
     finite : bool
         If True replace the infinite values in addition to the NaNs
     """
-    n = len(df)
+    n = len(data)
 
     if vars is None:
-        vars = df.columns.to_list()
+        vars = data.columns.to_list()
     else:
-        vars = df.columns.intersection(list(vars)).to_list()
+        vars = data.columns.intersection(list(vars)).to_list()
 
     if finite:
         lst = [np.inf, -np.inf]
         to_replace = {v: lst for v in vars}
-        df.replace(to_replace, np.nan, inplace=True)
+        data.replace(to_replace, np.nan, inplace=True)
         txt = "non-finite"
     else:
         txt = "missing"
 
-    df = df.dropna(subset=vars)
-    df.reset_index(drop=True, inplace=True)
-    if len(df) < n and not na_rm:
+    data = data.dropna(subset=vars)
+    data.reset_index(drop=True, inplace=True)
+    if len(data) < n and not na_rm:
         msg = "{} : Removed {} rows containing {} values."
-        warn(msg.format(name, n - len(df), txt), PlotnineWarning, stacklevel=3)
-    return df
+        warn(
+            msg.format(name, n - len(data), txt), PlotnineWarning, stacklevel=3
+        )
+    return data
 
 
 def to_rgba(colors, alpha):
@@ -781,7 +818,7 @@ class RegistryHierarchyMeta(type):
     def __init__(cls, name, bases, namespace):
         if not hasattr(cls, "_registry"):
             cls._registry = {}
-            cls._hierarchy = collections.defaultdict(list)
+            cls._hierarchy = defaultdict(list)
         else:
             cls._registry[name] = cls
             cls._hierarchy[name].append(name)
@@ -868,7 +905,7 @@ def copy_missing_columns(df, ref_df):
         idx = np.repeat(0, l1)
 
     for col in cols:
-        df[col] = ref_df.iloc[idx, _loc(col)].values
+        df[col] = ref_df.iloc[idx, _loc(col)].to_numpy()
 
 
 def data_mapping_as_kwargs(args, kwargs):
@@ -1073,7 +1110,7 @@ def cross_join(df1: pd.DataFrame, df2: pd.DataFrame) -> pd.DataFrame:
     all_columns = list(pd.Index(list(df1.columns) + list(df2.columns)))
     df1["key"] = 1
     df2["key"] = 1
-    return pd.merge(df1, df2, on="key").loc[:, all_columns]
+    return df1.merge(df2, on="key").loc[:, all_columns]
 
 
 def to_inches(value: float, units: str) -> float: